+++ uni_modules/vk-unicloud/vk_modules/vk-unicloud-page/libs/vk-unicloud/vk-unicloud-callFunctionUtil.js
@@ -747,6 +747,7 @@ class CallFunctionUtil {
 			sysErr = true;
 			errMsg = JSON.stringify(res);
 		}
+		if (!errMsg) errMsg = "";
 		if (errMsg.indexOf("timeout") > -1) {
 			sysErr = true;
 			errMsg = globalErrorCode["cloudfunction-timeout"] || "请求超时，请重试！";